Added the limit into the cache key.
Don't really like this as we can't reuse queries with different limits.
In theory a query with a smaller limit should be able to use a query
with a bigger limit. The issue is that results aren't sorted and we may
get back a different result set (particulalary over time).
But of course if we are worried about changes over time then we
shouldn't be using the cache.
Not certain what is the right answer so I've taken the conserverative
approach.