Made the cache list immutable.
Fixed a bug where a users of the cache could contaminate the cache by
changing the returned list of entities in the list.
Now return a copy of the cache values so the user can do what they want.

Cleaned out all of the place holder entities so its is obvious that they
still need to be implemented.